handled the preview of the single user by admin

+++ controller/ClientPostRequest/DeletePost.js
@@ -1,8 +1,17 @@
 const { ClientPostRequest } = require("../../model/ClientPosts");
-
+const { ClientSignUp } = require("../../model/ClientSignUp");
 module.exports.deletePostByClient = async (req, res) => {
   const { postId } = req.params;
-  const { _id } = req.client;
+  const { _id, role } = req.client;
+
+  const checkRole = ClientSignUp.find({ account_type: role });
+
+  if (!checkRole) {
+    return res.status(403).jsoN({
+      message: "access denied",
+      status: "error",
+    });
+  }
 
   const ClientPost = await ClientPostRequest.findByIdAndDelete(
     { _id: postId, clientId: _id },
@@ -13,12 +22,10 @@ module.exports.deletePostByClient = async (req, res) => {
           status: "error",
         });
       } else {
-        return res
-          .status(200)
-          .json({
-            message: "Deleting of post was successful",
-            status: "success",
-          });
+        return res.status(200).json({
+          message: "Deleting of post was successful",
+          status: "success",
+        });
       }
     }
   );
